+++ applications/fishing-map/src/features/workspace/common/InfoModal.tsx
@@ -1,6 +1,8 @@
 import { Fragment, useState, useCallback, useMemo } from 'react'
 import ReactHtmlParser from 'react-html-parser'
 import { useTranslation } from 'react-i18next'
+import { useSelector } from 'react-redux'
+import { uniqBy } from 'lodash'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import Tabs, { Tab } from '@globalfishingwatch/ui-components/dist/tabs'
 import Modal from '@globalfishingwatch/ui-components/dist/modal'
@@ -8,36 +10,60 @@ import { IconButton } from '@globalfishingwatch/ui-components'
 import { DatasetStatus } from '@globalfishingwatch/api-types/dist'
 import { getDatasetDescriptionTranslated } from 'features/i18n/utils'
 import { getDatasetLabel } from 'features/datasets/datasets.utils'
+import { isGFWUser } from 'features/user/user.slice'
 import styles from './InfoModal.module.css'
 
 type InfoModalProps = {
   dataview: UrlDataviewInstance
   onClick?: (e: React.MouseEvent) => void
   className?: string
+  onModalStateChange?: (open: boolean) => void
 }
 
-const InfoModal = ({ dataview, onClick, className }: InfoModalProps) => {
+const InfoModal = ({ dataview, onClick, className, onModalStateChange }: InfoModalProps) => {
   const { t } = useTranslation()
   const [modalInfoOpen, setModalInfoOpen] = useState(false)
+  const gfwUser = useSelector(isGFWUser)
   const dataset = dataview.datasets?.[0]
 
   const tabs = useMemo(() => {
-    return dataview.datasets?.flatMap((dataset) => {
+    const uniqDatasets = dataview.datasets ? uniqBy(dataview.datasets, (dataset) => dataset.id) : []
+    return uniqDatasets.flatMap((dataset) => {
       if (dataview.config?.datasets && !dataview.config?.datasets?.includes(dataset.id)) {
         return []
       }
       const description = getDatasetDescriptionTranslated(dataset)
+      const rawQueries = dataset.configuration?.documentation?.queries
+      const queries = Array.isArray(rawQueries) ? rawQueries : [rawQueries as unknown as string]
       return {
         id: dataset.id,
         title: getDatasetLabel(dataset),
         content: (
-          <p className={styles.content}>
-            {/**
-             * For security reasons, we are only parsing html
-             * coming from translated descriptions
-             **/}
-            {description.length > 0 ? ReactHtmlParser(description) : dataset.description}
-          </p>
+          <Fragment>
+            <p className={styles.content}>
+              {/**
+               * For security reasons, we are only parsing html
+               * coming from translated descriptions
+               **/}
+              {description.length > 0 ? ReactHtmlParser(description) : dataset.description}
+            </p>
+            {gfwUser && (
+              <div className={styles.content}>
+                <h2 className={styles.subtitle}>Queries used</h2>
+                {queries?.length ? (
+                  queries?.map((query: string, index: number) => (
+                    <div key={index}>
+                      <a target="_blank" href={query} rel="noreferrer">
+                        query {index + 1}
+                      </a>
+                    </div>
+                  ))
+                ) : (
+                  <p>none specified</p>
+                )}
+              </div>
+            )}
+          </Fragment>
         ),
       }
     })
@@ -49,15 +75,17 @@ const InfoModal = ({ dataview, onClick, className }: InfoModalProps) => {
   const handleClick = useCallback(
     (e: React.MouseEvent) => {
       setModalInfoOpen(true)
+      if (onModalStateChange) onModalStateChange(true)
       if (onClick) {
         onClick(e)
       }
     },
-    [onClick]
+    [onClick, onModalStateChange]
   )
   const onModalClose = useCallback(() => {
     setModalInfoOpen(false)
-  }, [])
+    if (onModalStateChange) onModalStateChange(false)
+  }, [onModalStateChange])
 
   const isSingleTab = tabs?.length === 1
 
@@ -114,6 +142,7 @@ const InfoModal = ({ dataview, onClick, className }: InfoModalProps) => {
               tabs={tabs}
               activeTab={activeTab?.id}
               onTabClick={(tab: Tab) => setActiveTab(tab)}
+              buttonSize="verybig"
             />
           )}
         </Modal>