+++ applications/vessel-history/src/features/event-filters/filters.hooks.ts
@@ -1,18 +1,11 @@
 import { useDispatch, useSelector } from 'react-redux'
-import {
-  selectFilters,
-  updateFilters,
-  availableEventFilters,
-} from './filters.slice'
+import { selectFilters, updateFilters, availableEventFilters } from './filters.slice'
 
 export const useApplyFiltersConnect = () => {
   const dispatch = useDispatch()
   const filters = useSelector(selectFilters)
 
-  const setFilter = (
-    filter: availableEventFilters,
-    value: boolean
-  ) => {
+  const setFilter = (filter: availableEventFilters, value: boolean) => {
     const newFilters = {
       ...filters,
       [filter]: value,
@@ -20,10 +13,7 @@ export const useApplyFiltersConnect = () => {
     dispatch(updateFilters(newFilters))
   }
 
-  const setDate = (
-    filter: 'start' | 'end',
-    value: string
-  ) => {
+  const setDate = (filter: 'start' | 'end', value?: string) => {
     const newFilters = {
       ...filters,
       [filter]: value,
@@ -38,6 +28,6 @@ export const useApplyFiltersConnect = () => {
   return {
     setFilter,
     setDate,
-    resetFilters
+    resetFilters,
   }
 }