+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.93
+
+### Patch Changes
+
+- Updated dependencies [5ca834a5]
+  - @globalfishingwatch/layer-composer@10.0.1
+
 ## 1.1.92
 
 ### Patch Changes