+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.63
+
+### Patch Changes
+
+- Updated dependencies [4f1d3fec]
+  - @globalfishingwatch/ui-components@12.3.0
+
+## 1.1.62
+
+### Patch Changes
+
+- Updated dependencies [5ca834a5]
+  - @globalfishingwatch/layer-composer@10.0.1
+
+## 1.1.61
+
+### Patch Changes
+
+- Updated dependencies [f3fa1b87]
+  - @globalfishingwatch/api-types@6.3.0
+
 ## 1.1.60
 
 ### Patch Changes